+++ modules/@apostrophecms/piece-type/ui/apos/components/AposDocsManager.vue
@@ -270,13 +270,13 @@ export default {
     },
     async onArchive(id) {
       const piece = this.findDocById(this.items, id);
-      if (await this.archive(this.options.action, id, !!piece.lastPublishedAt)) {
+      if (await this.archive(piece)) {
         apos.bus.$emit('content-changed');
       }
     },
     async onRestore(id) {
       const piece = this.findDocById(this.items, id);
-      if (await this.restore(this.options.action, id, !!piece.lastPublishedAt)) {
+      if (await this.restore(piece)) {
         apos.bus.$emit('content-changed');
       }
     },